backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.28
psycopg2-binary>=2.9
django-cors-headers>=4.3
pytest>=8.0
pytest-django>=4.8
